Show forest

Show forest plot with forestploter for fullfill the functionality of bregr results.

Usage

br_show_forest(breg, clean = TRUE, ...)

Arguments

breg

An object of class breg with results.

clean

If TRUE, remove "Group" or "Focal" variable column when the values are same, and reduce repeat values in column "Group", "Focal", and "Variable".

...

Arguments passing to forestploter::forest().

See also